merged with upstream (#2)
* Add badge from repology.org

* Make badge right aligned

* Highlight ruby docs as man pages

Neovim does not have a dedicated filetype for ruby docs but the man page
syntax seems to work reasonably well.

* Add tests for "cat-exec" mode

"cat-exec" mode is not a real "mode" that is documented for the end
user.  It is meant to describe the case when the nvimpager script just
calls `exec cat`.

* Use Ubuntu bionic on travis

The support for xenial was removed from the ppa repository.

* Bump version number to 0.7

* Use api to set mappings

This breaks compatibility with neovim < v0.4.

* Use the api to set options

This breaks compatibility with neovim < v0.4.

* Update required neovim version in README

* Fix "q" map in visual mode

* Fix lint errors

* Use table.concat() instead of self written join()

* Use PARENT as an env variable instead of PPID

PID is a special bash variable and is reset by every bash process.  On
NixOS many commands are wrapped in bash scripts to set up special
environments (especially neovim).  Thus the exported PPID is not visible
in neovim on NixOS.

Found with the help of Johan Myréen at
https://unix.stackexchange.com/questions/611627.

* Remove pending test marker for osx

* Use default background=dark in the tests

Upstream did change the default for 'background' from light to dark in
version 0.3.2.  In 991ccbfcb1a7df6cf4a94d21e85d029a95860464 I reset
background=light on the command line for all tests.

This commit changes all test fixtures to "dark" ansi escape sequences
and removes the setting from the test command lines.  This makes it
easier to write new test cases and cleans up the test code.

* Use busted's finally() to clean up local temp file

* Add failing tests for issue #28

* Ignore stdin if file argument is present

One has to specify "-" explicitly to still read stdin into a buffer if a
file argument is present.  This is what less does.

Closes #28.

* Update TODO list

* Start ansi highlighting functions in lua

This aims at fully replacing the AnsiEsc plugin with lua.

* Create one namespace for all manual highlights

* Compute a group name for generated highlighting groups

* Fix lua regex and set some local options

* Properly initialize parser state

* Skip zero length regions in renderer

* Fix line and column offset when highlighting

* Start defineing the highlight groups

* Initialize namespace var later so we do not need to mock it

* Add special split function for lua51 bug

* Fix tests

* Add tests for the split function

* Rename bright terminal attribute to statndout

* Move highlight command computation to a function

* Simplify iteration over terminal attributes

* Fix missing function argument

* Fix parser for multible matches on a line

* Implement reset for fg and bg colors

* Parse some additional term attributes

* Implement conceal escape sequence

* Implement some "switch off" escape sequences

* Clean up code

* Reuse cat mode cache variable for defined syntax groups

* Remove old code

* Add some small doc comments to functions

* Use nicer function in table syntax

This makes it look more like a local function definition in the module.

* Break long lines

* Store local function in cache variable

* Add more tests for parser function

* Start tokenizing function with tests

* Finish tokenizer with tests

* Add color computation functions

* Remove extra reloads of nvimpager script from tests

* Use helper function for repetitive tests

* Add special parser functions for special color sequences

* Fix tokenizer

* Fix main parser function

* Remove ansi esc plugin from makefile

* Update readme after ansiesc removal

* Remove ansiesc stuff from gitignore

* Fix names of highlight groups for hex colors

* Use cterm[fb]g argument to :highlight

* Adjust high pallet colors

* Bump version number to 0.8

* Add zsh completion

* Improve completion

Nvimpagers options do not work after files and the -h and -v options
terminate the script so other options make no sense with them.

* Install zsh completion with the makefile

* Add missing line continuation

* Set synmaxcol=0 and nowrap in ansi-esc mode

* Remove unused make variable

* Turn closure into proper function

The only variable that was bound to the closure is still available when
it is called.  Presumably having many closures defined and called is
less efficient than one function with an extra argument.

* Move test helper function to global scope

This makes it possible to split the long describe block.

* Rewrite tokenizer for ansi esc sequences

By splitting the strings early in the tokenizer a lot of later code
becomes obsolete.

* Use --clean in tests

* Use a newer macOS to fix build errors

https://travis-ci.community/t/the-default-osx-image-xcode9-4-is-now-unusable-because-macos-10-13-is-eol/10628

Co-authored-by: Lucas Hoffmann <lucc@posteo.de>
Co-authored-by: Lucas Hoffmann <lucc@users.noreply.github.com>